Added files via upload
I added a few Player variables to keep track of the shots taken, hits, etc for end of game percentages.  Also added a endGame() method to display the message and percentages.  Also finished the checkShot() method.  However, for testing this we need to have the ships populating their positions, so we need some GUI action.